+++ packages/cli/src/commands/testscript.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { Command } from "commander";
 import fs from "node:fs";
 import path from "node:path";
